+++ pkg/debugger/proxy.go
@@ -6,7 +6,6 @@ import (
 	"sort"
 	"time"
 
-	"github.com/openservicemesh/osm/pkg/catalog"
 	"github.com/openservicemesh/osm/pkg/certificate"
 )
 
@@ -60,7 +59,7 @@ func printProxies(w http.ResponseWriter, proxies map[certificate.CommonName]time
 }
 
 func (ds DebugConfig) getConfigDump(cn certificate.CommonName, w http.ResponseWriter) {
-	pod, err := catalog.GetPodFromCertificate(cn, ds.kubeController)
+	pod, err := ds.kubeController.GetPodFromCertificate(cn)
 	if err != nil {
 		log.Error().Err(err).Msgf("Error getting Pod from certificate with CN=%s", cn)
 	}
@@ -70,7 +69,7 @@ func (ds DebugConfig) getConfigDump(cn certificate.CommonName, w http.ResponseWr
 }
 
 func (ds DebugConfig) getProxy(cn certificate.CommonName, w http.ResponseWriter) {
-	pod, err := catalog.GetPodFromCertificate(cn, ds.kubeController)
+	pod, err := ds.kubeController.GetPodFromCertificate(cn)
 	if err != nil {
 		log.Error().Err(err).Msgf("Error getting Pod from certificate with CN=%s", cn)
 	}
