+++ tests/scenarios/traffic_split_with_apex_service_test.go
@@ -229,6 +229,6 @@ func getProxy(kubeClient kubernetes.Interface) (*envoy.Proxy, error) {
 
 	certCommonName := certificate.CommonName(fmt.Sprintf("%s.%s.%s", tests.ProxyUUID, tests.BookbuyerServiceAccountName, tests.Namespace))
 	certSerialNumber := certificate.SerialNumber("123456")
-	proxy := envoy.NewProxy(certCommonName, certSerialNumber, nil)
+	proxy := envoy.NewProxy(certCommonName, certSerialNumber, nil, nil)
 	return proxy, nil
 }
